cmake: Build and install CMake and PkgConfig files
In order to make it easy for other projects to depend on libalternatives,
we want them to be able to detect it in their build scripts to properly
depend on the library.

This introduces a CMake module script and a pkgconfig file to support
virtually all build tools that support dependencies by either of those
systems.

As part of this, we bump the minimum version to CMake 3.12 because
that version introduces most of the built-in modules for generating
CMake config modules.