Add basic Makefile for buidling without cmake
The Makefile is only for buidling the library and does not
include any unit tests. It's only useful if you plan on
building the library without cmake and cunit.

Unlike with cmake, some #defines need to be manually set with
CFLAGS. See CMakeLists.txt for details.

make -C src -f Makefile.gnu